requirements.txt
Flask
gunicorn
Jinja2
torch
torchvision
res-facenet
pillow